+++ api/cache/stats.js
@@ -1,2 +1,0 @@
-import handler from '../../routes/cache/stats.js';
-export default handler;