+++ api/deploy/status.js
@@ -1,2 +1,0 @@
-import handler from '../../routes/deploy/status.js';
-export default handler;