+++ api/server/metrics.js
@@ -1,2 +1,0 @@
-import handler from '../../routes/server/metrics.js';
-export default handler;