+++ api/server/tests.js
@@ -1,2 +1,0 @@
-import handler from '../../routes/server/tests.js';
-export default handler;