+++ api/v2/ready.js
@@ -1,2 +1,0 @@
-import handler from '../../routes/ready.js';
-export default handler;